backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy>=1.4
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
langchain==0.3.7
langchain-community==0.3.5
langgraph==0.2.39
ollama==0.4.2
redis==5.2.0
celery==5.4.0
chromadb==0.5.5
sentence-transformers==3.2.1
ctransformers==0.2.27
openai-whisper>=20240930
python-docx==1.1.2
reportlab==4.2.5
APScheduler==3.10.4
